First Commit
This is my first commit to this repository and branch. This has all the work on the combat that I've done so far.

At the moment it has entities, gear, prefabs, and the turn order. Next thing I'll be doing is having it where you can attack and have the enemies attack back.

After that I will go back in and add things like resistences, stat buffs, and abilities.